refactor: 애니메이션 설정 값 AnimationView nested로 변경
